updated model to int64 numbers for firestore compat

+++ internal/domain/models.go
@@ -14,11 +14,11 @@ const (
 // migration task document from Firestore
 type MigrationJob struct {
 	ID             string    `firestore:"-" json:"id"`
-	FromDevice     int32     `firestore:"fromDevice" json:"from_device"`
-	ToDevice       int32     `firestore:"toDevice" json:"to_device"`
+	FromDevice     int64     `firestore:"fromDevice" json:"from_device"`
+	ToDevice       int64     `firestore:"toDevice" json:"to_device"`
 	Status         JobStatus `firestore:"status" json:"status"`
-	FromDate       int32     `firestore:"fromDate,omitempty" json:"from_date"`
-	ToDate         int32     `firestore:"toDate,omitempty" json:"to_date"`
+	FromDate       int64     `firestore:"fromDate,omitempty" json:"from_date"`
+	ToDate         int64     `firestore:"toDate,omitempty" json:"to_date"`
 	ConvertData    bool      `firestore:"convertData" json:"convert_data"`
 	MigrateAllData bool      `firestore:"migrateAllData" json:"migrate_all_data"`
 	OverrideData   bool      `firestore:"createdBy" json:"override_data"`
